feat: Refactor ConfigurationMenu in two components
